Establish basic markup structure in GuestDetail
